refactor: Modal 컴포넌트 스타일 구조 개선
- constants.ts를 styles.ts로 통합하여 파일 구조 단순화
- Figma 노드 370:3126 기준 디자인 값 상세 문서화
- MODAL_STYLES 및 DEFAULT_CONFIG를 styles.ts에서 직접 export
- 모든 색상 값을 commons/constants/color.ts의 Colors 토큰 사용
